src/requirements.txt
mysql.connector==2.2.9
psycopg2==3.2.1
python-dotenv=1.0.1
